+++ package/src/renderer/Container.tsx
@@ -1,7 +1,6 @@
-import type { DependencyManager } from "../DependencyManager";
-import type { DrawingContext } from "../DrawingContext";
-
-import { Node } from "./Node";
+import type { DependencyManager } from "./DependencyManager";
+import type { DrawingContext } from "./DrawingContext";
+import { Node } from "./nodes";
 
 export class Container extends Node {
   redraw: () => void;
