+++ package/src/renderer/nodes/index.ts
@@ -1,4 +1,3 @@
 export * from "./Drawing";
 export * from "./Declaration";
-export * from "./Container";
 export * from "./Node";
